Adjust for new data schema imports. split camstim into camstim ephys and stim utils
