+++ src/aind_metadata_mapper/stimulus/__init__.py
@@ -1,0 +1,3 @@
+"""
+Stimulus utils
+"""
